+++ src/client/java/com/arcticquests/pills/SplashOminousClient.java
@@ -4,13 +4,13 @@ import com.arcticquests.raid.SplashOminous;
 
 import net.fabricmc.api.ClientModInitializer;
 import net.fabricmc.fabric.api.client.particle.v1.ParticleFactoryRegistry;
-import net.minecraft.client.particle.EndRodParticle;
+import net.minecraft.client.particle.EmotionParticle;
 
 public class SplashOminousClient implements ClientModInitializer {
  
     @Override
     public void onInitializeClient() {
-    ParticleFactoryRegistry.getInstance().register(SplashOminous.OMINOUS_PARTICLE, EndRodParticle.Factory::new);
+    ParticleFactoryRegistry.getInstance().register(SplashOminous.OMINOUS_PARTICLE, EmotionParticle.AngryVillagerFactory::new);
 
 }
 }
