+++ src/main/resources/assets/splash_ominous/lang/en_us.json
@@ -1,7 +1,8 @@
 
 {
-    "item.splash_ominous.splash_ominous_bottle" : "Splash Ominous Bottle",
-    "item.splash_ominous.lingering_ominous_bottle" : "Lingering Ominous Bottle",
+    "item.minecraft.potion.effect.ominous_potion" : "Splash Ominous Bottle",
+    "item.minecraft.potion.splash_bottle.effect.ominous_potion" : "Splash Ominous Bottle",
+    "item.minecraft.potion.lingering_bottle.effect.ominous_potion" : "Lingering Ominous Bottle",
 
     "entity.splash_ominous.ominous_particle": "Ominous Particles"
 }
