+++ src/main/resources/assets/splash_ominous/models/item/ominous_potion.json
@@ -1,0 +1,6 @@
+{
+    "parent": "item/generated",
+    "textures": {
+      "layer0": "splash_ominous:item/ominous_bottle"
+    }
+  }
