+++ src/main/resources/splash_ominous.mixins.json
@@ -3,7 +3,7 @@
 	"package": "com.arcticquests.raid.mixin",
 	"compatibilityLevel": "JAVA_21",
 	"mixins": [
-		"BrewingRecipeRegistryMixin1"
+
 	],
 	"injectors": {
 		"defaultRequire": 1
